bin file renamed to main
