+++ bin/main.rb
